✨ Improve base fsm
